Address review committee feedback (round 2)
Codex Round-2 must-fix #1 (vii°7 in minor): the harmonic-minor leading-
tone seventh chord is fully diminished (G#-B-D-F in A minor), not half-
diminished. Round-1's degree-raise fix gave the correct root but the
suffix selection was still using the natural-minor halfDiminished rule.
Fix: extend `suffixForNumeral` with a minor-key + degree-7 + diminished
quality branch returning `diminishedSeventh`. Major-key vii°7 keeps
half-diminished by natural-major spelling. Regression tests for
vii°7 in A minor and vii°7/ii in C major.

Codex Round-2 must-fix #2 (`applied: null` slips past truthiness check):
`fromJSON` validation used `value.applied ? recurse : undefined`, which
silently coerces `null` to undefined. Fix: validate by presence (`'applied'
in value`) plus an explicit `null`/non-object guard before recursing.
Tests for `applied: null`, `applied: 'V'` (string), and the
`applied: undefined` (treated same as absent) cases.

Reorganization: `src/roman-numeral.test.ts` exceeded the 500-line lint
limit (max-lines rule, skipBlankLines+skipComments) after adding the
round-1 regression tests. Split into:
  - `src/roman-numeral.test.ts` (parser + value-type tests, 375 lines)
  - `src/key-roman.test.ts` (Key-aware behavior tests, 218 lines)

This is the split testing-expert flagged in round 1 as a suggestion;
the lint rule converted it to a hard fail in round 2.

bun run validate green: 783 tests, 100% coverage, build clean.
